% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsignif-package.R
\docType{package}
\name{ggsignif-package}
\alias{ggsignif}
\alias{ggsignif-package}
\title{ggsignif: Significance Brackets for 'ggplot2'}
\description{
Enrich your 'ggplots' with group-wise comparisons. This package provides an easy way to indicate if two groups are significantly different. Commonly this is shown by a bracket on top connecting the groups of interest which itself is annotated with the level of significance (NS, *, **, ***). The package provides a single layer (geom_signif()) that takes the groups for comparison and the test (t.test(), wilcox.text() etc.) as arguments and adds the annotation to the plot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://const-ae.github.io/ggsignif/}
  \item \url{https://github.com/const-ae/ggsignif}
  \item Report bugs at \url{https://github.com/const-ae/ggsignif/issues}
}

}
\author{
\strong{Maintainer}: Constantin Ahlmann-Eltze \email{artjom31415@googlemail.com} (\href{https://orcid.org/0000-0002-3762-068X}{ORCID}) [contributor]

Authors:
\itemize{
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) [contributor]
}

}
\keyword{internal}
