% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_signif.R
\name{stat_signif}
\alias{stat_signif}
\alias{geom_signif}
\title{Create significance layer}
\usage{
stat_signif(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  comparisons = NULL,
  test = "wilcox.test",
  test.args = NULL,
  annotations = NULL,
  map_signif_level = FALSE,
  y_position = NULL,
  xmin = NULL,
  xmax = NULL,
  margin_top = 0.05,
  step_increase = 0,
  tip_length = 0.03,
  size = 0.5,
  textsize = 3.88,
  family = "",
  vjust = 0,
  parse = FALSE,
  manual = FALSE,
  orientation = NA,
  ...
)

geom_signif(
  mapping = NULL,
  data = NULL,
  stat = "signif",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  comparisons = NULL,
  test = "wilcox.test",
  test.args = NULL,
  annotations = NULL,
  map_signif_level = FALSE,
  y_position = NULL,
  xmin = NULL,
  xmax = NULL,
  margin_top = 0.05,
  step_increase = 0,
  extend_line = 0,
  tip_length = 0.03,
  size = 0.5,
  textsize = 3.88,
  family = "",
  vjust = 0,
  parse = FALSE,
  manual = FALSE,
  orientation = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are
either the names of 2 values on the x-axis or the 2 integers that
correspond to the index of the columns of interest.}

\item{test}{the name of the statistical test that is applied to the values of
the 2 columns (e.g. \code{t.test}, \code{wilcox.test} etc.). If you implement a
custom test make sure that it returns a list that has an entry called
\code{p.value}.}

\item{test.args}{additional arguments for the test method}

\item{annotations}{character vector with alternative annotations, if not null
test is ignored}

\item{map_signif_level}{Boolean value, if the p-value are directly written as
annotation or asterisks are used instead. Alternatively one can provide a
named numeric vector to create custom mappings from p-values to annotation:
For example: \code{c("***"=0.001, "**"=0.01, "*"=0.05)}.
Alternatively, one can provide a function that takes a numeric argument
(the p-value) and returns a string.}

\item{y_position}{numeric vector with the y positions of the brackets}

\item{xmin, xmax}{numeric vector with the positions of the left and right
sides of the brackets, respectively}

\item{margin_top}{numeric vector how much higher that the maximum value that
bars start as fraction of total height}

\item{step_increase}{numeric vector with the increase in fraction of total
height for every additional comparison to minimize overlap.}

\item{tip_length}{numeric vector with the fraction of total height that the
bar goes down to indicate the precise column}

\item{size}{change the width of the lines of the bracket}

\item{textsize}{change the size of the text}

\item{family}{change the font used for the text}

\item{vjust}{move the text up or down relative to the bracket}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{manual}{Boolean flag that indicates that the parameters are provided
with a data.frame. This option is necessary if one wants to plot different
annotations per facet.}

\item{orientation}{The orientation of the layer. The default (‘NA’)
automatically determines the orientation from the aesthetic mapping.
In the rare event that this fails it can be given explicitly by setting
'orientation' to either "x" or "y"}

\item{...}{other arguments passed on to \code{layer}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{extend_line}{Numeric that allows to shorten (negative values) or extend
(positive value) the horizontal line between groups for each comparison;
defaults to 0.}
}
\description{
Create significance layer
}
\examples{
\dontrun{
library(ggplot2)
library(ggsignif)

ggplot(mpg, aes(class, hwy)) +
  geom_boxplot() +
  geom_signif(comparisons = list(
    c("compact", "pickup"),
    c("subcompact", "suv")
  ))

ggplot(mpg, aes(class, hwy)) +
  geom_boxplot() +
  geom_signif(
    comparisons = list(
      c("compact", "pickup"),
      c("subcompact", "suv")
    ),
    map_signif_level = function(p) sprintf("p = \%.2g", p)
  )

ggplot(mpg, aes(class, hwy)) +
  geom_boxplot() +
  geom_signif(
    annotations = c("First", "Second"),
    y_position = c(30, 40), xmin = c(4, 1), xmax = c(5, 3)
  )
}

}
