% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_annotation.R
\name{stat_signif}
\alias{stat_signif}
\alias{geom_signif}
\title{Create significance layer}
\usage{
stat_signif(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  comparisons = NULL, test = "wilcox.test", test.args = NULL,
  annotations = NULL, map_signif_level = FALSE, y_position = NULL,
  margin_top = 0.05, step_increase = 0, tip_length = 0.03, ...)

geom_signif(mapping = NULL, data = NULL, stat = "signif",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, comparisons = NULL, test = "wilcox.test",
  test.args = NULL, annotations = NULL, map_signif_level = FALSE,
  y_position = NULL, margin_top = 0.05, step_increase = 0,
  tip_length = 0.03, ...)
}
\arguments{
\item{comparisons}{A list of length-2 vectors.
The entries in the vector are either the names of 2 values on the x-axis
or the 2 integers that correspond to the index of the columns of interest}

\item{test}{the name of the statistical test that is applied to the values of the 2 columns (e.g. t.test, wilcox.test etc.)}

\item{test.args}{additional arguments for the test method}

\item{annotations}{character vector with alternative annotations, if not null test is ignored}

\item{map_signif_level}{boolean value, if the p-value are directly written as annotation or asterisks are used instead.
Alternatively one can provide a named numeric vector to create custom mappings from p-values to annotation:
For example: c("***"=0.001, "**"=0.01, "*"=0.05)}

\item{y_position}{numeric vector with the precise locations of the bars}

\item{margin_top}{numeric vector how much higher that the maximum value that bars start as fraction of total height}

\item{step_increase}{numeric vector with the increase in fraction of total height for every additional comparison to
minimize overlap.}

\item{tip_length}{numeric vector with the fraction of total height that the bar goes down to indicate the precise column}
}
\description{
Create significance layer
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, hwy)) +
 geom_boxplot() +
 geom_signif(comparisons = list(c("compact", "pickup"), c("subcompact", "suv")))
}
